Feat: 필터링 초기화 기능 추가, 필터링 패널 컴포넌트로 분리
